scraper/requirements.txt
gspread
oauth2client
beautifulsoup4
requests
python-dotenv
python-dateutil
ruff
